requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
python-docx
PyPDF2
google-generativeai
python-dotenv
streamlit
numpy
plotly
pandas
colorama